% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choiceofT.R
\name{choiceofT}
\alias{choiceofT}
\title{Visualize EM fitting for each cell. BIC is used to choose optimal CNV group.}
\usage{
choiceofT(Y, gcfit.temp, gctemp, norm_index, offset, T, alpha,
  minCountQC = 20, filename)
}
\arguments{
\item{Y}{read depth matrix across all cells}

\item{gcfit.temp}{bin-specific GC content bias matrix across all single cells}

\item{gctemp}{a vector giving values of bin-specific GC content after \code{gcfit.tempj}
outlier removal (for better fitting)}

\item{norm_index}{indices of normal/diploid cells}

\item{offset}{offset matrix across all cells}

\item{T}{a vector of integers indicating number of CNV groups. Use BIC to select optimal
number of CNV groups. If \code{T = 1}, assume all reads are from normal regions so that
EM algorithm is not implemented. Otherwise, we assume there is always a CNV group of heterozygous
deletion and a group of null region. The rest groups are representative of different duplication
states.}

\item{alpha}{adjusted/absolute copy number matrix across all cells}

\item{minCountQC}{the minimum read coverage required for EM fitting. Defalut is \code{20}}

\item{filename}{the name of output pdf file}
}
\value{
pdf file with EM fitting results and two plots: log likelihood, and BIC versus the number of CNV groups.
}
\description{
Determine the number of CNV group via BIC for each. A pdf file containing EM fitting results and plots is generated.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
