% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coverageObj.scopeDemo}
\alias{coverageObj.scopeDemo}
\title{Pre-stored coverageObj.scope data for demonstration purposes}
\format{Pre-computed using whole genome sequencing data of
 three single cells from Navin et al.}
\usage{
coverageObj.scopeDemo
}
\description{
Pre-stored coverageObj.scope data for demonstration purposes
}
\references{
Navin, N., Kendall, J., Troge, J., Andrews, P., Rodgers, L.,
 McIndoo, J., Cook, K., Stepansky, A., Levy, D., Esposito, D.,
 et al. (2011). Tumour evolution inferred by single-cell
 sequencing. Nature 472, 90-94.
}
\keyword{datasets}
