% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep.Pois.R
\name{Estep.Pois}
\alias{Estep.Pois}
\title{Expectation step for SCOPE normalization}
\usage{
Estep.Pois(Yj, Nj, betatemp, fGCi, vec_pi, min.prop)
}
\arguments{
\item{Yj}{read depth vector for each single cell}

\item{Nj}{a numerirc value of total number of reads per cell}

\item{betatemp}{bin-specific bias estimated using negative control samples}

\item{fGCi}{estimated vector of GC content bias fitting from the M-step \code{\link{Mstep.Pois}}}

\item{vec_pi}{vector of incident rates for CNV events that span bin \emph{i}
from the M-step \code{\link{Mstep.Pois}}}

\item{min.prop}{the minmimum of mixture proportion for candidate CNV groups,
which serves as a stopping metric for EM algorithm}
}
\value{
A list with components
  \item{Z}{Matrix of optimized missing data to be fed into M-step \code{\link{Mstep.Pois}}}
  \item{obs_LL}{Observed log-Likelihood}
  \item{keep_going}{If normal proportion or a mixture proportion reached the minimum value,
   \code{keep_going = FALSE}, and output \code{NULL}, aka this case won't be involved in optimal number of CNV
   group selection based upon BIC.}
}
\description{
Expectation step of GC content fitting for SCOPE normalization.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
