% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estep.R
\name{Estep}
\alias{Estep}
\title{Expectation step for SCOPE normalization}
\usage{
Estep(fGCi, vec_pi, Yj, offsetj)
}
\arguments{
\item{fGCi}{estimated vector of GC content bias fitting from the M-step \code{\link{Mstep}}}

\item{vec_pi}{vector of incident rates for CNV events that span bin \emph{i}
from the M-step \code{\link{Mstep}}}

\item{Yj}{read depth vector for each single cell}

\item{offsetj}{vector of offset terms for each cell}
}
\value{
\item{Z}{Matrix of optimized missing data to be fed into M-step \code{\link{Mstep}}}
}
\description{
Expectation step of GC content fitting for SCOPE normalization.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
