% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGC.R
\name{fitGC}
\alias{fitGC}
\title{Fit a loess curve of f(GC)}
\usage{
fitGC(gctemp, gcfit.temp)
}
\arguments{
\item{gctemp}{a vector giving values of GC content for each bin after quality control}

\item{gcfit.temp}{vector of bin-specific GC content biases for each single cell}
}
\value{
\item{fGCi}{Estimated vector of GC content bias fitting}
}
\description{
Fit a loess curve of f(GC)
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
