% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcoverage.scDNA.R
\name{getcoverage.scDNA}
\alias{getcoverage.scDNA}
\title{Get read coverage from single-cell DNA sequencing}
\usage{
getcoverage.scDNA(bambedObj, mapqthres, mask.ref, seq)
}
\arguments{
\item{bambedObj}{object returned from \code{\link[CODEX2]{getbambed}}}

\item{mapqthres}{mapping quality threshold of reads}

\item{mask.ref}{a GRanges object indicating bad regions/bins, such as segmental
duplication regions and gaps near telomeres/centromeres, which need to be masked
prior to getting coverage}

\item{seq}{the sequencing method to be used. This should be either "paired-end"
or "single-end"}
}
\value{
\item{Y}{Read depth matrix}
}
\description{
Get read coverage for each genomic bin across all single cells
from scDNA-seq.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
