% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfGC.Pois.random.R
\name{getfGC.Pois.random}
\alias{getfGC.Pois.random}
\title{Get GC content bias fitting using Expectation-Maximization algorithm}
\usage{
getfGC.Pois.random(gcfitj, gctemp, Yj, Nj, betatemp, numGroup,
  verbose.plot = FALSE, gctemp.keep, min.prop)
}
\arguments{
\item{gcfitj}{vector of bin-specific GC content biases for each single cell}

\item{gctemp}{a vector giving values of bin-specific GC content after \code{gcfitj}
outlier removal (for better fitting)}

\item{Yj}{read depth vector for each single cell}

\item{Nj}{a numerirc value of total number of reads per cell}

\item{betatemp}{bin-specific bias estimated using negative control samples}

\item{numGroup}{a vector of integers indicating number of CNV groups. Use BIC to select optimal
number of CNV groups. If \code{numGroup = 1}, assume all reads are from normal regions so that
EM algorithm is not implemented. Otherwise, we assume there is always a CNV group of heterozygous
deletion and a group of null region. The rest groups are representative of different duplication
states.}

\item{verbose.plot}{logical, whether to plot GC content bias fitting results using EM and the
choice of optimal CNV group. Default is \code{FALSE}}

\item{gctemp.keep}{a vector giving values of bin-specific GC content without \code{gcfitj}
outlier removal. For better fitting, extreme GC content biases need to be excluded from EM input.
But a post hoc admission of all bins is necessary.}

\item{min.prop}{the minmimum of mixture proportion for candidate CNV groups,
which serves as a stopping metric for EM algorithm}
}
\value{
A list with components
  \item{BIC}{BIC for CNV group selection}
  \item{logL}{Observed log-Likelihood}
  \item{fGCi}{EM estimated vector of GC content bias fitting given \code{gctemp}}
  \item{Z}{Matrix of optimized missing data using EM algorithm}
  \item{vec_pi}{Vector of EM estimated incident rate for CNV events that span bin \emph{i}}
  \item{K}{Choice of optimal CNV group number based upon BIC}
  \item{fGCi.keep}{EM estimated vector of GC content bias fitting given \code{gctemp.keep}}
}
\description{
Fit a Poisson generalized linear model to normalize the raw read depth data from
 single-cell DNA sequencing, without latent factors under the case-control setting. SCOPE
 implements an EM algorithm with random initialization to unmask null regions. This is a
 bulit-in function within \code{\link{multi_run.Pois}}. We don't recommend running this function
 independently.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
