% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfGC.R
\name{getfGC}
\alias{getfGC}
\title{Get GC content bias fitting using Expectation-Maximization algorithm across all cells}
\usage{
getfGC(gcfit.temp, gctemp, Y, norm_index, offset, T, alpha, minCountQC)
}
\arguments{
\item{gcfit.temp}{bin-specific GC content bias matrix across all single cells}

\item{gctemp}{a vector giving values of bin-specific GC content after \code{gcfit.tempj}
outlier removal (for better fitting)}

\item{Y}{read depth matrix across all cells}

\item{norm_index}{indices of normal/diploid cells}

\item{offset}{offset matrix across all cells}

\item{T}{a vector of integers indicating number of CNV groups. Use BIC to select optimal
number of CNV groups. If \code{T = 1}, assume all reads are from normal regions so that
EM algorithm is not implemented. Otherwise, we assume there is always a CNV group of heterozygous
deletion and a group of null region. The rest groups are representative of different duplication
states.}

\item{alpha}{adjusted/absolute copy number matrix across all cells}

\item{minCountQC}{the minimum read coverage required for EM fitting}
}
\value{
A list with components
  \item{fGC.hat}{EM estimated GC content bias matrix given optimal CNV group selection}
  \item{alpha}{Absolute copy number matrix adjusted by EM fitting}
}
\description{
Fit a Poisson generalized linear model to normalize the raw read depth data from
 single-cell DNA sequencing, under the case-control setting. SCOPE
 implements an EM algorithm with random initialization to unmask null regions. This is a
 bulit-in function within \code{\link{normalize_scope}}.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
