% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfGCj.R
\name{getfGCj}
\alias{getfGCj}
\title{Get GC content bias fitting using Expectation-Maximization algorithm for each cell}
\usage{
getfGCj(gcfit.tempj, gctemp, Yj, offsetj, T, draw.plot = NULL, alphaj,
  minCountQC)
}
\arguments{
\item{gcfit.tempj}{vector of bin-specific GC content biases for each single cell}

\item{gctemp}{a vector giving values of bin-specific GC content after \code{gcfit.tempj}
outlier removal (for better fitting)}

\item{Yj}{read depth vector for each single cell}

\item{offsetj}{vector of offset terms for each cell}

\item{T}{a vector of integers indicating number of CNV groups. Use BIC to select optimal
number of CNV groups. If \code{T = 1}, assume all reads are from normal regions so that
EM algorithm is not implemented. Otherwise, we assume there is always a CNV group of heterozygous
deletion and a group of null region. The rest groups are representative of different duplication
states.}

\item{draw.plot}{logical, whether to plot GC content bias fitting results using EM and the
choice of optimal CNV group. Default is \code{NULL}}

\item{alphaj}{a vector giving values of adjusted/absolute copy number value for cell \emph{j}}

\item{minCountQC}{the minimum read coverage required for EM fitting}
}
\value{
A list with components
  \item{fGCi.obj}{A list with EM estimated vector of GC content bias fitting}
  \item{Z.obj = Z.obj}{A list with matrix of optimized missing data using EM algorithm}
  \item{vec_pi.obj}{A list with vector of EM estimated incident rate for CNV events that span bin \emph{i}}
  \item{bin.filter}{Indices of potential outlier bins}
  \item{loglik}{Observed log-Likelihood}
  \item{BIC}{BIC for CNV group selection}
}
\description{
Fit a Poisson generalized linear model to normalize the raw read depth data from
 single-cell DNA sequencing, under the case-control setting. SCOPE
 implements an EM algorithm with random initialization to unmask null regions. This is a
 bulit-in function within \code{\link{getfGC}}.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
