% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmapp.R
\name{getmapp}
\alias{getmapp}
\title{Compute mappability}
\usage{
getmapp(ref, genome = NULL)
}
\arguments{
\item{ref}{GRanges object returned from \code{\link[CODEX2]{getbambed}}}

\item{genome}{by default, \code{genome = BSgenome.Hsapiens.UCSC.hg19}.
To calculate mappability for hg38, specify \code{genome = BSgenome.Hsapiens.UCSC.hg38}}
}
\value{
\item{mapp}{Vector of mappability for each bin/target}
}
\description{
Compute mappability for each bin. Note that scDNA sequencing is
whole-genome amplification and the mappability score is essential to
determine variable binning method. Mappability track for 100-mers on the
GRCh37/hg19 human reference genome from ENCODE is pre-saved. Compute the mean
of mappability scores that overlapped reads map to bins, weighted by the width
of mappability tracks on the genome reference. Use liftOver utility to calculate
mappability for hg38, which is pre-saved as well.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
