% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsampQC.R
\name{getsampQC}
\alias{getsampQC}
\title{Get QC metrics for single cells}
\usage{
getsampQC(bambedObj)
}
\arguments{
\item{bambedObj}{object returned from \code{\link[CODEX2]{getbambed}}}
}
\value{
\item{QCmetric}{A matrix containing total number/proportion of reads,
   total number/proportion of mapped reads, total number/proportion of mapped
   non-duplicate reads, and number/proportion of reads with mapping quality
   greater than 20}
}
\description{
Perform QC step on single cells.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
