% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsumexp.R
\name{logsumexp}
\alias{logsumexp}
\title{Logarithm of summation of exponentials}
\usage{
logsumexp(xx)
}
\arguments{
\item{xx}{a numeric value or vector}
}
\value{
A numeric value giving natural logarithm of summation of exponentials
}
\description{
Computes the logarithm of summation of numeric exponentials.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
