% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapp_hg38}
\alias{mapp_hg38}
\title{GRanges with mappability scores for hg38}
\format{A GRanges object with 21584930 ranges and 1 metadata column of mappability scores}
\usage{
mapp_hg38
}
\description{
Use liftOver utility to convert hg19 coordinates to hg38
}
\references{
\url{http://rohsdb.cmb.usc.edu/GBshape/cgi-bin/hgFileUi?db=hg19&g=wgEncodeMapability}
}
\keyword{datasets}
