% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.Pois.R
\name{Mstep.Pois}
\alias{Mstep.Pois}
\title{Maximization step for SCOPE normalization}
\usage{
Mstep.Pois(Z, gcfitj, gctemp)
}
\arguments{
\item{Z}{matrix of optimized missing data from the E-step \code{\link{Estep.Pois}}}

\item{gcfitj}{vector of bin-specific GC content biases for each single cell}

\item{gctemp}{a vector giving values of GC content for each bin after quality control}
}
\value{
A list with components
  \item{vec_pi}{Vector of incident rate for CNV events that span bin \emph{i},
   which is to be fed into E-step \code{\link{Estep.Pois}}}
  \item{fGCi}{Estimated vector of GC content bias fitting}
}
\description{
Maximization step of GC content fitting for SCOPE normalization.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
