% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.R
\name{Mstep}
\alias{Mstep}
\title{Maximization step for SCOPE normalization}
\usage{
Mstep(Z, gcfit.tempj, gctemp)
}
\arguments{
\item{Z}{matrix of optimized missing data from the E-step \code{\link{Estep}}}

\item{gcfit.tempj}{vector of bin-specific GC content biases for each single cell}

\item{gctemp}{a vector giving values of GC content for each bin after quality control}
}
\value{
A list with components
  \item{vec_pi}{Vector of incident rate for CNV events that span bin \emph{i},
   which is to be fed into E-step \code{\link{Estep}}}
  \item{fGCi}{Estimated vector of GC content bias fitting}
}
\description{
Maximization step of GC content fitting for SCOPE normalization.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
