% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_run.Pois.R
\name{multi_run.Pois}
\alias{multi_run.Pois}
\title{Get the optimal GC content bias fitting using Expectation-Maximization algorithm}
\usage{
multi_run.Pois(gcfitj, gctemp, Yj, Nj, betatemp, numGroup, rerun,
  verbose.plot = FALSE, qc.thres = 5e-05, min.prop = 0.002)
}
\arguments{
\item{gcfitj}{vector of bin-specific GC content biases for each single cell}

\item{gctemp}{a vector giving values of bin-specific GC content after \code{gcfitj}
outlier removal (for better fitting)}

\item{Yj}{read depth vector for each single cell}

\item{Nj}{a numerirc value of total number of reads per cell}

\item{betatemp}{bin-specific bias estimated using negative control samples}

\item{numGroup}{a vector of integers indicating number of CNV groups. Use BIC to select optimal
number of CNV groups. If \code{numGroup = 1}, assume all reads are from normal regions so that
EM algorithm is not implemented. Otherwise, we assume there is always a CNV group of heterozygous
deletion and a group of null region. The rest groups are representative of different duplication
states.}

\item{rerun}{specify the number of running EM algorithm with random initialization}

\item{verbose.plot}{logical, whether to plot the optimal GC content bias fitting results using EM and the
choice of optimal CNV group. Default is \code{FALSE}}

\item{qc.thres}{the lower bound of bin-specific GC content bias threshold}

\item{min.prop}{the minmimum of mixture proportion for candidate CNV groups,
which serves as a stopping metric for EM algorithm}
}
\value{
A list with components for the optimal GC content bias fitting using EM
  \item{BIC}{BIC for CNV group selection}
  \item{logL}{Observed log-Likelihood}
  \item{fGCi}{EM estimated vector of GC content bias fitting given \code{gctemp}}
  \item{Z}{Matrix of optimized missing data using EM algorithm}
  \item{vec_pi}{Vector of EM estimated incident rate for CNV events that span bin \emph{i}}
  \item{K}{Choice of optimal CNV group number based upon BIC}
  \item{fGCi.keep}{EM estimated vector of GC content bias fitting given \code{gctemp.keep}}
}
\description{
SCOPE implements an EM algorithm with random initialization to unmask null regions.
 Adopt BIC to choose optimal GC content bias fitting among multiple runs of
 \code{\link{getfGC.Pois.random}}. This is a bulit-in function within
 \code{\link{normalize_codex2_ns_noK_EM_random}}.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
