% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_codex2_ns_noK.R
\name{normalize_codex2_ns_noK}
\alias{normalize_codex2_ns_noK}
\title{Normalization of read depth without latent factors under the case-control setting}
\usage{
normalize_codex2_ns_noK(Y_qc, gc_qc, K = 1, norm_index, N)
}
\arguments{
\item{Y_qc}{read depth matrix after quality control}

\item{gc_qc}{vector of GC content for each bin after quality control}

\item{K}{default is \code{K = 1}}

\item{norm_index}{indices of normal/diploid cells}

\item{N}{library size factor, which is computed from the genome-wide read depth data}
}
\value{
A list with components
  \item{Yhat}{A list of normalized read depth matrix}
  \item{fGC.hat}{A list of estimated GC content bias matrix}
  \item{beta.hat}{A list of estimated bin-specific bias vector}
}
\description{
Assuming that all reads are from diploid regions, fit a Poisson generalized
 linear model to normalize the raw read depth data from
 single-cell DNA sequencing, without latent factors under the case-control setting.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
