% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_codex2_ns_noK_EM_random.R
\name{normalize_codex2_ns_noK_EM_random}
\alias{normalize_codex2_ns_noK_EM_random}
\title{Normalization of read depth without latent factors using Expectation-Maximization algorithm
 under the case-control setting}
\usage{
normalize_codex2_ns_noK_EM_random(Y_qc, gc_qc, K = 1, norm_index, N,
  numGroup, qc.thres = 5e-05, min.prop = 0.002)
}
\arguments{
\item{Y_qc}{read depth matrix after quality control}

\item{gc_qc}{vector of GC content for each bin after quality control}

\item{K}{default is \code{K = 1}}

\item{norm_index}{indices of normal/diploid cells}

\item{N}{library size factor, which is computed from the genome-wide read depth data}

\item{numGroup}{a vector of integers indicating number of CNV groups. Use BIC to select optimal
number of CNV groups. If \code{numGroup = 1}, assume all reads are from normal regions so that
EM algorithm is not implemented. Otherwise, we assume there is always a CNV group of heterozygous
deletion and a group of null region. The rest groups are representative of different duplication
states.}

\item{qc.thres}{the lower bound of bin-specific GC content bias threshold}

\item{min.prop}{the minmimum of mixture proportion for candidate CNV groups,
which serves as a stopping metric for EM algorithm}
}
\value{
A list with components
  \item{Yhat}{A list of normalized read depth matrix with EM}
  \item{fGC.hat}{A list of EM estimated GC content bias matrix}
  \item{beta.hat}{A list of EM estimated bin-specific bias vector}
}
\description{
Fit a Poisson generalized linear model to normalize the raw read depth data from
 single-cell DNA sequencing, without latent factors under the case-control setting. Model GC
 content bias using an expectation-maximization algorithm, which accounts for the different
 copy number states.
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
