% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreEst_Ploidy.R
\name{PreEst_ploidy}
\alias{PreEst_ploidy}
\title{Ploidy pre-initialization}
\usage{
PreEst_ploidy(Y, Yhat, ref, maxPloidy = 6, minPloidy = 1.5,
  minBinWidth = 5)
}
\arguments{
\item{Y}{raw read depth matrix after quality control procedure}

\item{Yhat}{normalized read depth matrix}

\item{ref}{GRanges object after quality control procedure}

\item{maxPloidy}{maximum ploidy candidate. Defalut is \code{6}}

\item{minPloidy}{minimum ploidy candidate. Defalut is \code{1.5}}

\item{minBinWidth}{the minimum number of bins for a changed segment. Defalut is \code{5}}
}
\value{
\item{ploidy.SoS}{Vector of pre-estimated ploidies for each cell}
}
\description{
Pre-estimate ploidies across all cells
}
\author{
Rujin Wang \email{rujin@email.unc.edu}
}
